Many pointless changes, also some worthwhile stuff.

+++ Platform.h
@@ -2,12 +2,13 @@
 #define HG_PLATFORM_H
 #include "TimeDirection.h"
 #include "SortedByIndex.h"
+#include <cstring> //for size_t -- TODO find proper header
 namespace hg
 {
     class Platform
     {
     public:
-        Platform(int x, int y, int xspeed, int yspeed, int width, int height, int index, TimeDirection timeDirection, int pauseLevel);
+        Platform(int x, int y, int xspeed, int yspeed, int width, int height, size_t index, TimeDirection timeDirection, int pauseLevel);
         Platform(const Platform& other, TimeDirection timeDirection, int pauseLevel);
         Platform(const Platform& other);
         Platform& operator=(const Platform& other);
@@ -18,7 +19,7 @@ namespace hg
         int getYspeed() const {return yspeed_;}
         int getWidth() const {return width_;}
         int getHeight() const {return height_;}
-        int getIndex() const {return index_;}
+        size_t getIndex() const {return index_;}
         TimeDirection getTimeDirection() const {return timeDirection_;}
         int getPauseLevel() const {return pauseLevel_;}
 
@@ -34,7 +35,7 @@ namespace hg
         int yspeed_;
         int width_;
         int height_;
-        int index_;
+        size_t index_;
         TimeDirection timeDirection_;
         int pauseLevel_;
     };
