+++ TimelineState.h
@@ -15,7 +15,7 @@ public:
     /**************
      * Constructs a timeline state of length timeLength containing no arrivals or departures.
      */
-	TimelineState(unsigned int timelineLength);
+	TimelineState(size_t timelineLength);
 
     /********************
      * Updates the timeline with new departures and returns the set of frames
